// client/src/utils/routsConsts.ts
export enum PublicConsts {
    LOGIN_ROUTE = "/login",
    REGISTER_ROUTE = "/register",
    BASKET_ROUTE = "/basket",
    HOME_ROUTE = '/'
  }
  
export enum AdminConsts {
    ADD_PICTURE_ROUTE = "/admin/addPicture",
    PICTURE_ARRAY_ROUTE = "/admin/pictureArray",
    EDIT_PICTURE_ROUTE = "/admin/editPicture",
}

export enum CategoryConsts {
    MODULAR_ROUTE = "/modular_paintings",
    FRAMED_ROUTE = "/framed_paintings",
    POSTERS_ROUTE = "/posters",
    CLOCK_ROUTE = "/clock_paintings",
    LIGHTBOXES_ROUTE = "/lightboxes",
}